course_service/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.0
Pillow==9.4.0
requests==2.31.0
